style changes, bug fix in setting threshold

+++ code/bt_sensor.py
@@ -140,7 +140,7 @@ class BluetoothReader(QObject):
         command = {'tx':f"set threshold {int(hpa)}", 'rx':'threshold'}
         msg = self.send_receive_command(command)
         try:
-            received_p = int(msg[1])
+            received_p = float(msg[1])
         except:
             received_p = 0
         if msg[0] == command['rx'] and received_p == int(hpa):
